requirements.txt
discord.py==2.2.2
pynput==1.7.6
